fixed several printf format warnings
git-svn-id: file:///svn/uhexen2/code/trunk@2084 eaf649e6-6c13-0410-b194-84221022858e

+++ hexen2/win_stuff/snd_win.c
@@ -364,7 +364,7 @@ static sndinitstat SNDDMA_InitDirect (void)
 	IDirectSoundBuffer_Play(pDSBuf, 0, 0, DSBPLAY_LOOPING);
 
 	if (snd_firsttime)
-		Con_SafePrintf ("%d bytes in sound buffer\n", dsbcaps.dwBufferBytes);
+		Con_SafePrintf ("%lu bytes in sound buffer\n", (unsigned long)dsbcaps.dwBufferBytes);
 
 	gSndBufSize = dsbcaps.dwBufferBytes;
